feat(todo): implement Todo module with components for task management
- Add TodoTab as the main entry point for the Todo module.
- Create TodoHeader component to display task statistics.
- Implement TodoInputBar for adding new tasks with priority and size selection.
- Develop TodoItem component for individual task display with expandable details.
- Introduce TodoList to manage and render the list of tasks.
- Add TodoSubItem for managing subtasks within a TodoItem.
- Establish configuration for priorities and sizes in todoConfig.
- Create useTodos hook for state management and persistence of tasks.
- Define types for Todo module in todoTypes.
- Implement utility functions for formatting and normalizing tasks in todoUtils.
